Initial version of CI for RSSingle
Hi @tanrax!

As part of our new collaboration, I've made an initial version for
automated CI for RSSingle.

It needs macOS to work first, so I've commented out a draft version of
the `pyinstaller` aspect, so I'll make this PR a draft for now when we
can discuss further, preferably in a GH issue for transparency (unless
you want it private?), but this workflow tests on PRs and pushes to any
branch, with flake8 across multiple Python versions. The pyinstaller
aspect only runs if the flake8 check passes.

What do you think?

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>