Add Docker support for RSSingle, and a workflow for builds
This commit adds a Dockerfile, .dockerignore, and a GH Action for
pushing container images to - and this is by default, other registries
can be used - Docker Hub (you need to generate a token and install it in
the repo), and GitHub Registry, which doesn't require a manual token.

The workflow won't build until #3 is merged, as it relies on successful
builds from that workflow before a container image is pushed. Once we use unit
tests, this will ensure container images aren't broken when pushed to a
remote registry.

I have marked this PR as draft for that reason.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>